+++ examples/mobile-native-example/android/app/src/main/java/com/elit/mobileexample/MainActivity.kt
@@ -1,0 +1,20 @@
+package com.elit.mobileexample
+
+import android.annotation.SuppressLint
+import android.os.Bundle
+import android.webkit.WebSettings
+import android.webkit.WebView
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+  @SuppressLint("SetJavaScriptEnabled")
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    val webView = WebView(this)
+    val settings: WebSettings = webView.settings
+    settings.javaScriptEnabled = true
+    settings.domStorageEnabled = true
+    webView.loadUrl("file:///android_asset/public/index.html")
+    setContentView(webView)
+  }
+}